refactor: 개수에 대한 값 int로 수정

+++ backend/src/main/java/com/yagubogu/checkin/dto/TeamCheckInCountResponse.java
@@ -3,6 +3,6 @@ package com.yagubogu.checkin.dto;
 public record TeamCheckInCountResponse(
         long id,
         String name,
-        long count
+        int count
 ) {
 }
